+++ include/Triangulo.hpp
@@ -1,15 +1,15 @@
 class Triangulo
 {
 private:
-    int resultado;
-    int m[3][3];
+    float resultado;
+    float m[3][3];
 
 public:
-    Triangulo(int _matriz[3][3]);
-    int getResultado() const;
+    Triangulo(float _matriz[3][3]);
+    float getResultado() const;
 
 private:
-    int sentidoPositivo();
-    int sentidoNegativo();
-    void setMatriz(int _matriz[3][3]);
+    float sentidoPositivo();
+    float sentidoNegativo();
+    void setMatriz(float _matriz[3][3]);
 };
